Streamlit/CIST-FASS/requirements.txt
folium
streamlit_folium
turfpy
boto3
matplotlib
seaborn
japanize-matplotlib
plotly
streamlit-pandas
streamlit-aggrid
requests